refactor(core): split project.rs into project/ and project/task/
Separate Project and Task aggregate roots into distinct modules.
Project operations remain in project/mod.rs while task CRUD operations
move to project/task/mod.rs.
